+++ changelog.md
@@ -1,6 +1,7 @@
 v0.0.4-alpha
 ------------
 * Added Shield of the Sacred Oath, a ritually empowered shield that reduces all damage you take when blocking
+* Fixed bug in aligned wand core mana regen
 
 v0.0.3-alpha
 ------------